MAINT: Break out AlignmentStats
