+++ src/nw_align.h
@@ -1,11 +1,10 @@
 #ifndef _ALIGN_H_
 #define _ALIGN_H_
 
-#include <cstddef>
-#include <string>
-
 class FASTASequence;
 class AlignStats;
+
+#include "alignment_stats.h"
 
 class NeedlemanWunschAligner
 {
@@ -17,23 +16,6 @@ public:
 
 private:
     // TODO: Make alignment matrix and gap penalties configurable
-};
-
-class AlignStats
-{
-public:
-    std::string queryId;
-    std::string subjectId;
-    double pctIdentity;
-    int alignmentLength;
-    int mismatches;
-    int gapOpenings;
-    size_t queryStart;
-    size_t queryEnd;
-    size_t subjectStart;
-    size_t subjectEnd;
-    double EValue;
-    int bitScore;
 };
 
 #endif // _ALIGN_H_
